refactor: [DP-1127] линтинг и логи
