+++ package.json
@@ -6,7 +6,7 @@
     "dev": "next dev --turbopack ",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "prepare": "husky ",
     "proto:gen": "rimraf src/shared/proto && proto-gen.sh",
     "build:proto": "rimraf src/shared/proto && mkdirp src/shared/proto && globstar -- grpc_tools_node_protoc --experimental_allow_proto3_optional --ts_proto_out=./src/shared/proto --ts_proto_opt=useExactTypes=false,outputTypeRegistry=false,outputServices=generic-definitions,outputServices=nice-grpc,stringEnums=true,oneof=unions-value,esModuleInterop=true --descriptor_set_out=./src/shared/proto/protoset.bin --include_imports -I ./proto \"./@(proto_deps|proto)/**/*.proto\"",
