+++ src/entities/tariff/model/use-get-tariffs.ts
@@ -2,7 +2,7 @@
 
 import { keepPreviousData, useQuery } from "@tanstack/react-query";
 import { useTariffsClient } from "@/shared/grpc/clients/use-tariffs-client";
-import { QUERY_KEYS } from "@/shared/lib/query-keys";
+import { QUERY_KEYS } from "@/shared/lib";
 import { GetTariffsRequest } from "@/shared/proto/api_tips_tariff/v1/api_tips_tariff";
 
 export const useGetTariffs = (req: GetTariffsRequest) => {
