+++ src/features/manage-user/model/use-get-user.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { useAccessClient } from "@/shared/grpc/clients/use-user-client";
-import { QUERY_KEYS } from "@/shared/lib/query-keys";
+import { QUERY_KEYS } from "@/shared/lib";
 import { GetUserRequest } from "@/shared/proto/api_tips_access/v1/api_tips_access";
 
 export const useGetUser = (req: Partial<GetUserRequest>) => {
