+++ src/features/manage-user/ui/user-form.tsx
@@ -25,7 +25,6 @@ export const UserForm: FC<UserFormProps> = ({
   const { t } = useTranslations();
   const userQuery = useGetUser({ userId });
 
-  console.log(userQuery);
   const {
     register,
     handleSubmit,
@@ -54,8 +53,6 @@ export const UserForm: FC<UserFormProps> = ({
   });
 
   const onSubmit = (data: UserFormValues) => {
-    console.log(data);
-    console.log(userId);
     onSuccess(data);
   };
