+++ src/features/user-registry-filters/ui/user-registry-filters.tsx
@@ -15,7 +15,7 @@ import { ListFilterIcon } from "lucide-react";
 import { useRouter, useSearchParams } from "next/navigation";
 import { FC, useEffect } from "react";
 import { Controller, useForm } from "react-hook-form";
-import { createFilterMapper } from "@/shared/lib/create-filter-mapper";
+import { createFilterMapper } from "@/shared/lib";
 import { useTranslations } from "@/shared/locale/translations";
 import { GetUsersRequest_Filter } from "@/shared/proto/api_tips_access/v1/api_tips_access";
