+++ src/shared/grpc/GrpcClientProvider.tsx
@@ -7,7 +7,7 @@ import { AuthMiddleware } from "./AuthMiddleware";
 import { GrpcClientsContextValue, grpcClientsContext } from "./context";
 import { errorMiddleware } from "./errorMiddleware";
 import { loggerMiddleware } from "./loggerMiddleware";
-import { TokenService } from "../lib/tokenService";
+import { TokenService } from "../lib";
 
 const channel = createChannel(
   process.env.NEXT_PUBLIC_API_URL ?? "https://beta.api-tips.api.quckoo.net",
