+++ src/shared/lib/index.ts
@@ -7,3 +7,7 @@ export {
   subtractionDecimal,
   sumDecimal,
 } from "./decimal";
+export { default as dayjs } from "./dayjs-in";
+export { TokenService } from "./tokenService";
+export { createFilterMapper } from "./create-filter-mapper";
+export { QUERY_KEYS } from "./query-keys";
